_archive_optimized_20251116_020206/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etf-rotation"
version = "0.1.0"
description = "精简版ETF轮动流程，涵盖因子生成与直接因子级WFO"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "ETF Rotation Team", email = "quant@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "numba>=0.58",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=23.11",
    "isort>=5.12",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
